Typescript added in IPC and main
