+++ app/main/helpers/ace/colorComments.js
