+++ app/main/helpers/os.js
