+++ app/main/helpers/paths.js
