+++ app/main/helpers/terminal.js
