+++ app/main/ipc/misc.ts
@@ -1,0 +1,5 @@
+import { ipcMain, IpcMainInvokeEvent, shell } from "electron";
+
+ipcMain.handle("open-in-browser", (_: IpcMainInvokeEvent, url: string) => {
+    shell.openExternal(url);
+});
