+++ helpers/debuggerWindow/debuggerWindow.js
@@ -2,8 +2,8 @@ const { BrowserWindow, nativeImage, app, ipcMain } = require("electron")
 const path = require("path")
 const bus = require("../eventBus.js")
 
-const { getAppIcon } = require("../../app/helpers/requests.js")
-const { ASSETS_PATH } = require("../../app/helpers/paths.js")
+const { getAppIcon } = require("../../app/main/helpers/requests.js")
+const { ASSETS_PATH } = require("../../app/main/helpers/paths.js")
 
 async function createDebuggerWindow(mainWindow, title = "Debugger") {
     const overlayIconPath = path.join(ASSETS_PATH, "media", "debugger_icon.png")
