+++ package.json
@@ -3,12 +3,16 @@
     "version": "0.0.2-dev",
     "description": "Web-First Code Editor",
     "author": "Yurba.Developers",
-    "main": "main.js",
+    "main": "app/dist/main.js",
     "scripts": {
-        "start": "electron ."
+        "build": "tsc",
+        "start": "npm run build && electron ."
     },
     "devDependencies": {
-        "electron": "^39.2.3"
+        "@types/electron": "^1.4.38",
+        "@types/node": "^25.9.1",
+        "electron": "^39.2.3",
+        "typescript": "^6.0.3"
     },
     "dependencies": {
         "@xterm/addon-fit": "^0.11.0",
@@ -18,7 +22,6 @@
         "error-stack-parser": "^2.1.4",
         "flashot": "^1.4.1",
         "node-global-key-listener": "^0.3.0",
-        "typescript": "^6.0.3",
         "ws": "^8.19.0"
     }
 }
