+++ tsconfig.json
@@ -1,0 +1,24 @@
+{
+    "compilerOptions": {
+        "target": "ES2022",
+        "module": "CommonJS",
+        "outDir": "app/dist",
+        "rootDir": "app/main",
+        "strict": true,
+        "types": ["node", "electron"],
+        "lib": [
+            "ES2022"
+        ],
+        "esModuleInterop": true,
+        "skipLibCheck": true,
+        "allowJs": true
+    },
+    "include": ["app/main/**/*"],
+    "exclude": [
+        "ace/**/*",
+        "assets/**/*",
+        "helpers/**/*",
+        "node_modules",
+        "dist"
+    ]
+}
